move query to utils
